requirements.txt
pip==21.3.1
requests==2.26.0
GitPython==3.1.24
toml==0.10.2
flask==2.0.2
lxml==4.7.1
nebula2-python==2.6.0
pyhash==0.9.3
